ptrace: Add support for GETREGSET
Hello there,

While working with libunwind using the ptrace interface on linux, we faced an unexpected problem. Everything works fine for old architectures like arm or x86, but that's not the case any longer for more younger ones like aarch64.
The point is, it looks like linux kernel maintainers seem to not be willing to implement some interfaces for the most recent architectures. Among the interface that are not expected to be implemented, there are the following:
-       PTRACE_PEEKUSER
-       PTRACE_POKEUSER

These interfaces have been replace by a more generic one:
-       PTRACE_GETREGSET
-       PTRACE_SETREGSET

Therefore I propose the following patch. The idea is to use the new interface mechanism as far as possible. Please note that I only implemented it for the architectures that I could test, at least for the basics:
-       x86
-       x86_64
-       arm
-       aarch64

The patch is based on the libunwind adaption for BSD systems, and the gdb implementation.

Best regards

Frederic Berat
>From ee43b80f6c34bb5e1cb09699d7d2cfd0c12949a1 Mon Sep 17 00:00:00 2001
From: Frederic Berat <fberat@de.adit-jv.com>
Date: Thu, 12 May 2016 15:49:34 +0200
Subject: [PATCH] ptrace: Add support for GETREGSET

Adding support for PTRACE_GETREGSET that is expected to replace
PTRACE_PEEK for newer architectures.

Signed-off-by: Frederic Berat <fberat@de.adit-jv.com>